pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "pulse-counter-diabatic"
description = "Counter-diabatic pulse design for Rydberg atom arrays"
readme = "README.md"
version = "0.1.0"
authors = [
    { name = "Mauro Mendizabal", email="mauro.mendizabal-pico@pasqal.com" },
    { name = "Kemal Bidzhiev", email="kemal.bidzhiev@pasqal.com" },
    { name = "Elie Merhej", email="elie.merhej@pasqal.com" },
    { name = "Anton Quelle", email = "anton.quelle@pasqal.com"},
]
requires-python = ">=3.10"
license = {file = "LICENSE"}
classifiers=[
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
    "torch>=2.9.0",
    "pulser-core[torch]>=1.7.0",
    "emu_base"

]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "pre-commit",
    "emu_mps",
    "emu_sv"
]


[tool.isort]
line_length = 90
combine_as_imports = true
balanced_wrapping = true
lines_after_imports = 2
include_trailing_comma = true
multi_line_output = 5

[flake8]
max-line-length = 90
